# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Playbook": {
        "properties": {
          "citation_quality": {
            "default": 0.0,
            "title": "Citation Quality",
            "type": "number"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "grounding_score": {
            "default": 0.0,
            "title": "Grounding Score",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "llm_model": {
            "default": "claude-sonnet-4-6",
            "title": "Llm Model",
            "type": "string"
          },
          "llm_provider": {
            "default": "anthropic",
            "title": "Llm Provider",
            "type": "string"
          },
          "steps": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PlaybookStep"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "title": "Playbook",
        "type": "object"
      },
      "PlaybookStep": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "citation_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Citation Ids",
            "type": "array"
          },
          "order": {
            "title": "Order",
            "type": "integer"
          },
          "phase": {
            "enum": [
              "containment",
              "eradication",
              "recovery",
              "lessons_learned"
            ],
            "title": "Phase",
            "type": "string"
          }
        },
        "required": [
          "order",
          "phase",
          "action"
        ],
        "title": "PlaybookStep",
        "type": "object"
      },
      "ReviewAction": {
        "properties": {
          "action": {
            "enum": [
              "approve",
              "reject",
              "edit_and_approve"
            ],
            "title": "Action",
            "type": "string"
          },
          "edited_playbook": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Playbook"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "action"
        ],
        "title": "ReviewAction",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Agentic Cybersecurity Threat Intelligence & Incident Response Advisor",
    "title": "ACTIRA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/review/queue": {
      "get": {
        "operationId": "review_queue_api_review_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Queue",
        "tags": [
          "review"
        ]
      }
    },
    "/api/review/{incident_id}": {
      "post": {
        "description": "Atomic HiTL action \u2014 only one reviewer wins if concurrent requests race.",
        "operationId": "review_incident_api_review__incident_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Incident",
        "tags": [
          "review"
        ]
      }
    },
...
    "/api/v1/review/queue": {
      "get": {
        "operationId": "review_queue_api_v1_review_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Queue",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/review/{incident_id}": {
      "post": {
        "description": "Atomic HiTL action \u2014 only one reviewer wins if concurrent requests race.",
        "operationId": "review_incident_api_v1_review__incident_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Incident",
        "tags": [
          "review"
        ]
      }
    },
...
  }
}